# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "3&7 Training Platform API",
    "description": "API for the 3&7 Training and Recovery Platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/training-sessions/": {
      "post": {
        "tags": [
          "Training Sessions"
        ],
        "summary": "Create Training Session",
        "description": "Create a new training session.\n\nOnly coaches and admins can create training sessions.",
        "operationId": "create_training_session_training_sessions__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingSessionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingSession"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Training Sessions"
        ],
        "summary": "Get Training Sessions",
        "description": "Get all training sessions",
        "operationId": "get_training_sessions_training_sessions__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/training-sessions/{session_id}": {
      "get": {
        "tags": [
          "Training Sessions"
        ],
        "summary": "Get Training Session",
        "description": "Get a specific training session by ID",
        "operationId": "get_training_session_training_sessions__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Training Sessions"
        ],
        "summary": "Update Training Session",
        "description": "Update an existing training session.\n\nOnly coaches who created the session and admins can update training sessions.",
        "operationId": "update_training_session_training_sessions__session_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingSessionCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingSession"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Training Sessions"
        ],
        "summary": "Delete Training Session",
        "description": "Delete a training session.\n\nOnly coaches who created the session and admins can delete training sessions.",
        "operationId": "delete_training_session_training_sessions__session_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get Users",
        "description": "Get all users",
        "operationId": "get_users_api_users__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/training-sessions/{session_id}": {
      "get": {
        "tags": [
          "Training Sessions"
        ],
        "summary": "Get Training Session Details",
        "operationId": "get_training_session_details_api_training_sessions__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Training Session Details Api Training Sessions  Session Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SessionType": {
        "type": "string",
        "enum": [
          "strength",
          "conditioning",
          "recovery",
          "skills",
          "team_practice",
          "other"
        ],
        "title": "SessionType",
        "description": "Types of training sessions"
      },
...
      "TrainingPhase": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "duration_minutes": {
            "type": "integer",
            "title": "Duration Minutes"
          },
          "exercises": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Exercises"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "duration_minutes",
          "exercises"
        ],
        "title": "TrainingPhase",
        "description": "Model for a training phase (warm-up, main, cool-down)"
      },
...
      "TrainingSession": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "type": {
            "$ref": "#/components/schemas/SessionType"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "title": "End Time"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "max_participants": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Participants"
          },
          "phases": {
            "items": {
              "$ref": "#/components/schemas/TrainingPhase"
            },
            "type": "array",
            "title": "Phases"
          },
          "equipment_needed": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equipment Needed"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "coach_id": {
            "type": "string",
            "title": "Coach Id"
          },
          "coach_name": {
            "type": "string",
            "title": "Coach Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "title",
          "type",
          "date",
          "start_time",
          "end_time",
          "location",
          "phases",
          "id",
          "coach_id",
          "coach_name",
          "created_at",
          "updated_at"
        ],
        "title": "TrainingSession",
        "description": "Model for a training session with ID"
      },
      "TrainingSessionCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "type": {
            "$ref": "#/components/schemas/SessionType"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "title": "End Time"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "max_participants": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Participants"
          },
          "phases": {
            "items": {
              "$ref": "#/components/schemas/TrainingPhase"
            },
            "type": "array",
            "title": "Phases"
          },
          "equipment_needed": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equipment Needed"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "athlete_group_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Athlete Group Ids"
          },
          "individual_athlete_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Individual Athlete Ids"
          }
        },
        "type": "object",
        "required": [
          "title",
          "type",
          "date",
          "start_time",
          "end_time",
          "location",
          "phases"
        ],
        "title": "TrainingSessionCreate",
        "description": "Model for creating a training session"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/auth/token"
          }
        }
      }
    }
  }
}